Add error checking for getting Obsidian download URL from github.

+++ run_onchange_20_mint_setup.sh
@@ -6,8 +6,12 @@ sudo apt install remmina filelight gnome-tweaks qbittorrent flameshot glances -y
 
 ############################
 # OBSIDIAN LATEST
-URL=$(wget -q -O - https://api.github.com/repos/obsidianmd/obsidian-releases/releases/latest | grep 'deb"$' | awk -F'"' ' {print $4}  ')
-wget -q -O obsidian.deb "$URL"
+OBSIDIAN_URL=$(wget -q -O - https://api.github.com/repos/obsidianmd/obsidian-releases/releases/latest | grep 'deb"$' | awk -F'"' ' {print $4}  ')
+if [ -z "$OBSIDIAN_URL" ]; then
+  echo "Failed to fetch Obsidian URL"
+  exit 1
+fi
+wget -q -O obsidian.deb "$OBSIDIAN_URL"
 sudo dpkg -i obsidian.deb
 rm obsidian.deb
